👔 skip option 추가
