+++ packages/publint/src/types/global.d.ts
@@ -1,0 +1,3 @@
+declare type Booleanize<T> = {
+    [K in keyof T]: T[K] extends object ? Booleanize<T[K]> : boolean
+}
